chore(deps): Update package-lock.json with peer dependencies
This commit updates the `package-lock.json` file to align with the format used by newer versions of npm.

Running `npm install` with npm v7 or later automatically adds the `"peer": true"` property to packages that are identified as peer dependencies. This change reflects that update, ensuring consistency and proper dependency resolution without altering the application's functionality.